+++ sdwan_configuration_groups.tf
@@ -144,6 +144,9 @@ locals {
       try(profile.dhcp_servers, null) == null ? [] : [for dhcp_server in try(profile.dhcp_servers, []) : [
         sdwan_service_dhcp_server_feature.service_dhcp_server_feature["${profile.name}-${dhcp_server.name}"].version
       ]],
+      try(profile.dual_router_ha_features, null) == null ? [] : [for dual_router_ha_feature in try(profile.dual_router_ha_features, []) : [
+        sdwan_service_dual_router_ha_feature.service_dual_router_ha_feature["${profile.name}-${dual_router_ha_feature.name}"].version
+      ]],
       try(profile.eigrp_features, null) == null ? [] : [for eigrp_feature in try(profile.eigrp_features, []) : [
         sdwan_service_routing_eigrp_feature.service_routing_eigrp_feature["${profile.name}-${eigrp_feature.name}"].version
       ]],